test deserialization.
Here we run into "Fun" in the Dwarf Fortress sense of the word:

Go's type system just *isn't very good*.

Ideally we want to declare a bunch of `const( ErrAsdf ThePkgErrorCategory = "whatever"; [...])` stuff, and have switch statements on `switch errcat.Category(err) {` with `case ErrAsdf:` in that switch, right?

So that's easy enough and WeGotThis in plain ol go code.  *Slight* hankyness, because your constructor funcs and `Category()` take and return `interface{}` wildcards -- there's no way to declare `func Errorf(category Generic[kind=string]) error`, so if we want to accept people's typedef'd types, we're stuck going full on `interface{}`.  Similarly, we return an `interface{}` from `Category(error) interface{}` and that lets us switch on the result directly (it doesn't check you at compile time if you compare to some other type that "shouldn't" be here, but that's simply not possible without declaring a full type and set of helper funcs per set of category enums, which... we're not doing).

Now when we get to round-tripping serialization formats, are we any better off?

No.

Deserializing into our concrete error type will result in a `string`-kind with no additional type info -- because how would it be otherwise?

And these type-less strings are **not equal** to your typed string constants you may have declared in your API packages.

(No, setting `e2.Category_ = ErrorCategory("")` as a hint before unmarshal does not help if the `Category_` field is still `interface{}`: string kinds are not themselves settable, so stdlib json.Unmarshal will still replace that value with a less-typeinfo'd one.)

So: if you want to deserialize these errors, and have a typed category on them: declare your own concrete type for receiving the deserialization.  Copypaste the necessary interface methods.  It'll work.  The good news is: it's only 10 lines of ugly.

Signed-off-by: Eric Myhre <hash@exultant.us>