Propose a new method for amending context to errors; also try out using text templating.
The core idea here is to stick such a thing in a defer at the top of
the public functions of a package (much like how the RequireErrorHasCategory
function is used), and use it to make sure certain context appears in the
details, or the message, or both.  For example, you might use it to always
make sure your current "job ID" (insert appropriate application-level term here)
is always listed in the error message.

The templating is because I notice that when I use errcat in practice,
I very frequently find myself repeating things in an Sprintf and in
the details map.  And I'd both A) rather not and B) rather be forced
to include things in the details pairs if they're that important in
the first place.

Hopefully this helps reliably produce meaningful error messages while
requiring much less repeating yourself with contextual information.

---

Another idea that floated around during this design is that perhaps we'd
rather compose *templates* right up until the last second.

This would let us provide functions for formatting right at the top,
where we actually know what kind of outputting we want to do; even
better, it would let us format them *twice*, i.e. once with color
for printf logs, and once as json/plain for API serialization.

It's possible to do some interesting stuff by doing 'Parse' onsite,
and replacing the funcmap later.  So, conceivably, we could make a pattern of
using some 'quote'-like functions that denote type of thinger, then replace
those placeholder functions with nicer renderers later, like ANSI terminal
colorized output.

The main problem with all this, of course, is that it would require
*storing templates*.  As cool as the results might be, doing that would
go against the grain of our basic composability requirements -- how
would we serialize them, and make sure other programs on the other side
of an API boundary could easily handle errors the same way we can?
That way lies madness (and an attempt to standardize templating systems
across languages); let's not.

Signed-off-by: Eric Myhre <hash@exultant.us>